refactor: move constant `defaultBuildTimeout` to `constants.ts`
