+++ core/test/integ/src/plugins/kubernetes/helm/config.ts
@@ -13,11 +13,12 @@ import { cloneDeep, omit } from "lodash"
 import { expectError, TestGarden } from "../../../../../helpers"
 import { PluginContext } from "../../../../../../src/plugin-context"
 import { dedent } from "../../../../../../src/util/string"
-import { defaultBuildTimeout, ModuleConfig } from "../../../../../../src/config/module"
+import { ModuleConfig } from "../../../../../../src/config/module"
 import { apply } from "json-merge-patch"
 import { getHelmTestGarden } from "./common"
 import { defaultHelmTimeout } from "../../../../../../src/plugins/kubernetes/helm/module-config"
 import stripAnsi = require("strip-ansi")
+import { DEFAULT_API_VERSION, defaultBuildTimeout } from "../../../../../../src/constants"
 import { PREVIOUS_API_VERSION } from "../../../../../../src/constants"
 
 describe("configureHelmModule", () => {
